Changed script to be able to use multiple files in context menu
As entries in the context menu are not able to receive more than one
argument contrary to entries in the "SendTo" folder I added a script to
be called from the context menu which will run for x selected files.

The script will then write the file paths to a temporary file and
decide if itself should call the main process.