Changed way of locking other instances
The previous system of locking future instances from the first call
involved creating a file after checking if it already exists and ending
the script if the file has been there already. This left a few issues
when the script broke. Now the file checks if the file is writable and
locks the file for the process time.